Реализована смена эмодзи при клике;верный формат даты с помощью moment
