+++ src/view/smart.js
@@ -1,0 +1,36 @@
+import AbstractView from "./abstract.js";
+
+export default class Smart extends AbstractView {
+  constructor() {
+    super();
+    this._data = {};
+  }
+  updateData(update) {
+    if (!update) {
+      return;
+    }
+
+    this._data = Object.assign(
+        {},
+        this._data,
+        update
+    );
+
+    this.updateElement();
+  }
+  updateElement() {
+    let prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+    prevElement = null;
+
+    this.restoreHandlers();
+  }
+  restoreHandlers() {
+    throw new Error(`Abstract method not implemented: resetHandlers`);
+  }
+}
